% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_data.R
\name{describe_data}
\alias{describe_data}
\title{Describe a Dataframe}
\usage{
describe_data(data)
}
\arguments{
\item{data}{A dataframe that you want to describe.}
}
\value{
This function does not return a value but prints detailed information about the dataframe.
}
\description{
Provides a comprehensive overview of a dataframe, including its structure,
summary statistics, the first few rows, column names, and the number of rows and columns.
This function is useful for getting a quick and detailed summary of a dataframe's contents and structure.
}
\examples{
# Assuming 'Education' is a dataframe
describe_data(Education)
}
