% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_logistic_regression.R
\name{fit_logistic_regression}
\alias{fit_logistic_regression}
\title{Fit Logistic Regression Model}
\usage{
fit_logistic_regression(x, y, learning_rate = 0.01, max_iter = 1000)
}
\arguments{
\item{x}{Matrix of predictors.}

\item{y}{Response variable.}

\item{learning_rate}{Learning rate for the gradient descent.}

\item{max_iter}{Maximum number of iterations.}
}
\value{
Model coefficients.
}
\description{
Fits a logistic regression model using Gradient Descent.
}
\examples{
# Generate a synthetic dataset
set.seed(123)
n <- 100  # number of samples
predictor variables
x <- matrix(rnorm(n * 2), ncol = 2)
coefficients
beta <- c(-1, 0.5, 1)
predictors
z <- cbind(1, x) \%*\% beta
logistic function which can be generate by my function logistic function as well
pr <- 1 / (1 + exp(-z))
response variable
y <- rbinom(n, 1, pr)
# Fit the logistic regression model
coefficients <- fit_logistic_regression(x, y)
print(coefficients)

}
