% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_logistic_regression_to_data.R
\name{fit_logistic_regression_to_data}
\alias{fit_logistic_regression_to_data}
\title{Fit Logistic Regression Model to Dataset}
\usage{
fit_logistic_regression_to_data(data, response_var, predictor_vars)
}
\arguments{
\item{data}{A dataframe containing the dataset.}

\item{response_var}{The name of the response variable in the dataset. The response variable should be binary (0 and 1).}

\item{predictor_vars}{A vector of names of predictor variables in the dataset.}
}
\value{
A numeric vector of coefficients from the logistic regression model, including the intercept.
}
\description{
This function fits a logistic regression model to a given dataset. It extracts the specified response variable and predictor variables from the dataset, converts them into the appropriate format, and then uses logistic regression function(in package as well) to compute the model coefficients.
}
\examples{
# 'Education' is a dataset with a binary response variable named 'response'
# and the first four columns are predictor variables
coefficients <- fit_logistic_regression_to_data(Education, "response", names(my_data)[1:4])
print(coefficients)
}
