% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{fn_visualize}
\alias{fn_visualize}
\title{Visualizations & variable transformation}
\usage{
fn_visualize(df, target, var, ln_trnfm = 0)
}
\arguments{
\item{df}{dataframe which target and variable present.}

\item{target}{Target variable.}

\item{var}{Variable need to consider.}

\item{ln_tranfm}{optional variable.}
}
\value{
Graphs.
}
\description{
Produce graph for the variable
}
\examples{
data(CTDF)
View(CTDF)
fn_visualize(df = CTDF,target = "TARGET",var = "HOLDING_PERIOD")
fn_visualize(df = CTDF,target = "TARGET",var = "NO_OF_L_CR_TXNS")
}
\author{
Rajesh Jakhotia
}
