% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{GOFit}
\alias{GOFit}
\title{Chi Sq - Goodness of Fit}
\usage{
GOFit(df, target, probability, g = 10)
}
\arguments{
\item{df}{dataframe which target and variable present.}

\item{target}{Target variable.}

\item{probability}{Predicted probabilities.}
}
\value{
Chi Sq. values
}
\description{
Checking Chi Sq - Goodness of Fit.
}
\examples{
data(CTDF)
View(CTDF)
mylogit<-glm(TARGET~HOLDING_PERIOD+NO_OF_L_CR_TXNS+AGE_BKT,data=CTDF,family = "binomial")
summary(mylogit)
vif(mylogit)
CTDF$prob<-predict.glm(mylogit,CTDF,type="response")
GOFit(CTDF,"TARGET","prob")
}
\author{
Rajesh Jakhotia
}
