% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello Function}
\usage{
hello(x)
}
\arguments{
\item{x}{any parameter here.}
}
\value{
"Hello, this is my function"
}
\description{
This function prints a greeting message. It's a simple demonstration of a function
that takes no arguments and returns no value, but performs an action (printing a message).
}
\examples{
hello() # prints "Hello, this is my function"

}
