% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logistic.R
\name{Logreg}
\alias{Logreg}
\title{Binomial Logistic regression}
\usage{
Logreg(X, y, maxit = 10000)
}
\arguments{
\item{X}{input matrix, of dimension n (sample number) by p (variable number);
each row is an observation vector.}

\item{y}{response variable.
Since we aim at binomial regression, it should be a vector with two levels (like a vector of 0 & 1, or a & b)}

\item{maxit}{the Maximum number of iterations when we use optimization to estimate the parameeter;
default is 10000. And the algorithm will stop according to its stopping criterion.}
}
\value{
A list containing the relevant data of regression result.
}
\description{
Fit binomial logistic regression via unpenalized maximum likelihood.
Use tuning-free optimization to get the result.
}
\examples{

### install and library package
#devtools::install_github("hobbitish1028/Logistic")
library(Logistic)

### Generate training data
sigma<-4
set.seed(123)
n<-1e4
p<-1e2
mu1<-rnorm(p)
mu2<-rnorm(p)
X1<-matrix(mu1+rnorm(n*p,0,sigma),n,p,byrow = TRUE)
X2<-matrix(mu2+rnorm(n*p,0,sigma),n,p,byrow = TRUE)

### Train data
X<-rbind(X1,X2)
y<-rep(c(1,0),each=n)

### Fit model
fit<-Logreg(X,y)

### accuracy of train data
fit$accuracy

### check the convergence
plot(fit$loss)

### get the prediction of train data
result<-fit$prediction

### get the probability of each sample belonging to two category
result<-fit$P

### get the parameter of the logistic regression (including intercept)
result<-fit$x

}
