% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize Specified Columns in a Dataset}
\usage{
standardize(data, columns)
}
\arguments{
\item{data}{A dataframe containing the dataset to be standardized.}

\item{columns}{A vector of strings specifying the names of the columns to be standardized.}
}
\value{
The original dataframe with specified columns standardized.
}
\description{
This function standardizes specified columns in a dataset. Standardization is performed
by subtracting the mean and dividing by the standard deviation for each value in the column.
This process transforms the data in each specified column to have a mean of 0 and a standard
deviation of 1, which is often useful for data analysis and machine learning.
}
\examples{
# Assuming 'EPL' is dataset in this package with columns: 'points', 'goals_scored', 'goals_conceded'
columns_to_standardize <- c("points", "goals_scored", "goals_conceded")
standardized_EPL <- standardize(EPL, columns_to_standardize)
head(standardized_EPL)
}
