\name{Logistic-package}
\alias{Logistic-package}
\alias{Logistic}
\docType{package}
\title{
  Binomial Logistic Regression
}
\description{
  Use logistic regression model to predict binary output (like 0 & 1, a & b)
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Chang Wang (wangchan@umich.edu)

Maintainer: Chang Wang <wangchan@umich.edu>
}
\keyword{ package }
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
