% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsampling.R
\name{upsampling}
\alias{upsampling}
\title{Upsample Imbalanced Data}
\usage{
upsampling(data, class_col)
}
\arguments{
\item{data}{A dataframe containing the dataset with a binary classification target.}

\item{class_col}{A string specifying the name of the column in `data` that contains the binary class labels.}
}
\value{
A dataframe where the number of instances in both classes of the target variable are balanced.
        The minority class is upsampled to match the number of instances in the majority class.
}
\description{
Increases the number of instances in the under-represented class in a binary classification dataset.
The function identifies the minority class and upsamples it (by sampling with replacement) to match
the size of the majority class. This is typically used to handle imbalanced datasets in classification problems.
}
\examples{
# 'Education' is a dataset with an imbalanced binary target column named 'response'

balanced_data <- upsampling(Education, "response")
head(balanced_data)

# Now 'balanced_data' will have an equal number of instances for both classes in 'response' column
}
